requirements.txt
aiofiles==23.2.1
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
bidict==0.22.1
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
cryptography==41.0.4
ecdsa==0.18.0
exceptiongroup==1.1.3
fastapi==0.103.1
fastapi-microsoft-identity==0.1.6
fastapi-socketio==0.0.10
frozenlist==1.4.0
google-api-core==2.12.0
google-api-python-client==2.102.0
google-auth==2.23.2
google-auth-httplib2==0.1.1
googleapis-common-protos==1.60.0
h11==0.14.0
httpcore==0.17.3
httplib2==0.22.0
httptools==0.6.0
httpx==0.24.1
idna==3.4
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
markdown2==2.4.10
MarkupSafe==2.1.3
multidict==6.0.4
nicegui==1.3.16
orjson==3.9.7
packaging==23.1
pluggy==1.3.0
protobuf==4.24.4
pscript==0.7.7
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.3.0
pydantic-settings==2.0.3
pydantic_core==2.6.3
Pygments==2.16.1
PyJWT==2.8.0
pyparsing==3.1.1
pytest==7.4.2
python-dotenv==1.0.0
python-engineio==4.7.1
python-jose==3.3.0
python-multipart==0.0.6
python-socketio==5.9.0
PyYAML==6.0.1
requests==2.31.0
rsa==4.9
simple-websocket==0.10.1
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
tomli==2.0.1
typing_extensions==4.8.0
uritemplate==4.1.1
urllib3==2.0.4
uvicorn==0.22.0
uvloop==0.17.0
vbuild==0.8.2
watchfiles==0.20.0
websockets==11.0.3
wsproto==1.2.0
yarl==1.9.2
